[py] moved get_path method to SeleniumManager class
