+++ py/selenium/webdriver/wpewebkit/webdriver.py
@@ -18,7 +18,7 @@
 import http.client as http_client
 
 from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
-from selenium.webdriver.common.driver_finder import DriverFinder
+from selenium.webdriver.common.selenium_manager import SeleniumManager
 from selenium.webdriver.remote.webdriver import WebDriver as RemoteWebDriver
 
 from .options import Options
@@ -56,7 +56,7 @@ class WebDriver(RemoteWebDriver):
             options = Options()
 
         self.service = Service(executable_path, port=port, log_path=service_log_path)
-        self.service.path = DriverFinder.get_path(self.service, options)
+        self.service.path = SeleniumManager.get_path(self.service, options)
         self.service.start()
 
         super().__init__(command_executor=self.service.service_url, desired_capabilities=desired_capabilities)
